Added response and request entities

+++ backend/src/main/java/com/example/tripguide/controller/CategoryController.java
@@ -1,7 +1,10 @@
 package com.example.tripguide.controller;
 
+import com.example.tripguide.controller.mapper.CategoryMapper;
 import com.example.tripguide.exception.BadRequestException;
 import com.example.tripguide.model.Category;
+import com.example.tripguide.payload.response.CategoryBasicResponse;
+import com.example.tripguide.payload.request.CategoryBasicRequest;
 import com.example.tripguide.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -11,7 +14,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -21,45 +23,62 @@ public class CategoryController {
     @Autowired
     private CategoryRepository categoryRepository;
 
+    private CategoryMapper categoryMapper = new CategoryMapper();
+
     @GetMapping("/categories")
-    public Page<Category> getAllCategories(Pageable pageable){
-        return this.categoryRepository.findAll(pageable);
+    public Page<CategoryBasicResponse> getAllCategories(Pageable pageable){
+        Page<Category> pageCategory = this.categoryRepository.findAll(pageable);
+        return pageCategory.map(this.categoryMapper::entityToBasicResponse);
     }
 
+    // Доделать, чтобы возвращал не одну категорию, а список
     @GetMapping("/category/")
-    public ResponseEntity<?> getCategory(String name) {
+    public ResponseEntity<?> getCategoryByName(@RequestParam String name) {
         Optional<Category> category = this.categoryRepository.findByName(name);
-        return category.map(response -> ResponseEntity.ok().body(response))
+        return category.map(response -> ResponseEntity.ok().body(this.categoryMapper.entityToBasicResponse(response)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/category/{id}")
-    public ResponseEntity<?> getCategory(@PathVariable Long id) {
+    public ResponseEntity<?> getCategoryById(@PathVariable Long id) {
         Optional<Category> category = this.categoryRepository.findById(id);
-        return category.map(response -> ResponseEntity.ok().body(response))
+        return category.map(response -> ResponseEntity.ok().body(this.categoryMapper.entityToBasicResponse(response)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @PostMapping("/category")
-    public ResponseEntity<Category> createCategory(@RequestBody Category category) throws URISyntaxException {
-        if (this.categoryRepository.existsByName(category.getName())) {
-            throw new BadRequestException("Такая категория уже создана");
+    public ResponseEntity<CategoryBasicResponse> createCategory(@RequestBody CategoryBasicRequest request)
+            throws URISyntaxException {
+        if (this.categoryRepository.existsByName(request.getName())) {
+            throw new BadRequestException("Такая категория уже создана!");
         }
 
+        Category category = this.categoryMapper.basicRequestToEntity(request);
         Category result = this.categoryRepository.save(category);
+        CategoryBasicResponse response = this.categoryMapper.entityToBasicResponse(result);
 
         return ResponseEntity.created(new URI("/api/category/" + result.getId()))
-                .body(result);
+                .body(response);
     }
 
     @PutMapping("/category/{id}")
-    public ResponseEntity<Category> updateCategory(@RequestBody Category category) {
+    public ResponseEntity<CategoryBasicResponse> updateCategory(@PathVariable Long id, @RequestBody CategoryBasicRequest request) {
+        if (this.categoryRepository.findById(id).isEmpty()) {
+            throw new BadRequestException("Такой категории не существует!");
+        }
+
+        Category category = this.categoryMapper.basicRequestToEntity(request);
         Category result = this.categoryRepository.save(category);
-        return ResponseEntity.ok().body(result);
+        CategoryBasicResponse response = this.categoryMapper.entityToBasicResponse(result);
+        return ResponseEntity.ok().body(response);
     }
 
     @DeleteMapping("/category/{id}")
     public ResponseEntity<?> deleteCategory(@PathVariable Long id) {
+        if (this.categoryRepository.findById(id).isEmpty()) {
+            throw new BadRequestException("Такой категории не существует!");
+        }
+
         this.categoryRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
